% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsc_assuntos.R
\docType{data}
\name{tjsc_assuntos}
\alias{tjsc_assuntos}
\title{tjsc_assuntos}
\format{
dataframe com 3614 obs e três colunas:
\describe{
  \item{\code{area}}{character área conforme tabela do TJSC}
  \item{\code{assunto}}{character assunto com código}
  \item{\code{assunto_}}{character assunto sem código, sem acento, minúsculo}
}
}
\source{
\url{https://tjsc.jus.br/}
}
\usage{
tjsc_assuntos
}
\description{
Assuntos do TJSC
}
\keyword{datasets}
