% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsc_baixar_cpopg.R
\name{tjsc_baixar_cpopg}
\alias{tjsc_baixar_cpopg}
\title{Baixar consulta processual de primeiro grau}
\usage{
tjsc_baixar_cpopg(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo (cnj)}

\item{diretorio}{Diretório onde serão armazenados os htmls}
}
\value{
html com dados processuais
}
\description{
Baixar consulta processual de primeiro grau
}
