% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsc_cpopg_baixar_tabela_docs.R
\name{tjsc_cpopg_baixar_tabela_docs}
\alias{tjsc_cpopg_baixar_tabela_docs}
\title{Baixa tabela de documentos dos processos do TJSC}
\usage{
tjsc_cpopg_baixar_tabela_docs(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo}

\item{diretorio}{Diretório onde armazenar as tabelas}
}
\value{
html
}
\description{
Baixa tabela de documentos dos processos do TJSC
}
